+++ packages/tools/kadena-cli/src/networks/utils/networkHelpers.ts
@@ -1,8 +1,8 @@
 import {
   defaultNetworksPath,
   defaultNetworksSettingsFilePath,
+  getNetworkFiles,
   networkDefaults,
-  networkFiles,
 } from '../../constants/networks.js';
 
 import {
@@ -16,6 +16,7 @@ import yaml from 'js-yaml';
 import path from 'path';
 import { z } from 'zod';
 import { services } from '../../services/index.js';
+import { KadenaError } from '../../services/service-error.js';
 
 export interface ICustomNetworkChoice {
   value: string; // | typeof skipSymbol | typeof createSymbol;
@@ -56,12 +57,14 @@ const networkSchema = z.object({
  * @returns {void} - No return value; the function writes directly to a file.
  */
 export async function writeNetworks(
+  kadenaDir: string,
   options: INetworkCreateOptions,
 ): Promise<void> {
   const { network } = options;
   const sanitizedNetwork = sanitizeFilename(network).toLowerCase();
   const networkFilePath = path.join(
-    defaultNetworksPath,
+    kadenaDir,
+    'networks',
     `${sanitizedNetwork}.yaml`,
   );
 
@@ -104,6 +107,9 @@ export async function writeNetworks(
 export async function removeNetwork(
   options: Pick<INetworkCreateOptions, 'network'>,
 ): Promise<void> {
+  if (defaultNetworksPath === null) {
+    throw new KadenaError('no_kadena_directory');
+  }
   const { network } = options;
   const sanitizedNetwork = sanitizeFilename(network).toLowerCase();
   const networkFilePath = path.join(
@@ -115,12 +121,21 @@ export async function removeNetwork(
 }
 
 export async function removeDefaultNetwork(): Promise<void> {
-  await services.filesystem.deleteFile(defaultNetworksSettingsFilePath);
+  if (defaultNetworksSettingsFilePath === null) {
+    throw new KadenaError('no_kadena_directory');
+  }
+
+  if (await services.filesystem.fileExists(defaultNetworksSettingsFilePath)) {
+    await services.filesystem.deleteFile(defaultNetworksSettingsFilePath);
+  }
 }
 
 export async function loadNetworkConfig(
   network: string,
 ): Promise<INetworksCreateOptions> {
+  if (defaultNetworksPath === null) {
+    throw new KadenaError('no_kadena_directory');
+  }
   const networkFilePath = path.join(defaultNetworksPath, `${network}.yaml`);
 
   const file = await services.filesystem.readFile(networkFilePath);
@@ -131,12 +146,17 @@ export async function loadNetworkConfig(
   return yaml.load(file) as INetworksCreateOptions;
 }
 
-export async function ensureNetworksConfiguration(): Promise<void> {
-  await services.filesystem.ensureDirectoryExists(defaultNetworksPath);
+export async function ensureNetworksConfiguration(
+  kadenaDir: string,
+): Promise<void> {
+  await services.filesystem.ensureDirectoryExists(
+    path.join(kadenaDir, 'networks'),
+  );
 
+  const networkFiles = getNetworkFiles(kadenaDir);
   for (const [network, filePath] of Object.entries(networkFiles)) {
     if (!(await services.filesystem.fileExists(filePath))) {
-      await writeNetworks(networkDefaults[network]);
+      await writeNetworks(kadenaDir, networkDefaults[network]);
     }
   }
 }